FInished Refactoring Adapter to Plugins and removing CastleWindsor
